Added cache optimization
Also learned that, on x86 machines, real 80 bits intermediate result is
the way to rock.

+++ src/comet/sma.d
@@ -7,7 +7,7 @@ import deimos.containers.tree;
 
 import std.algorithm;
 
-alias float Cost;
+alias double Cost;
 
 struct StateInfo {
   size_t count = 0;
